refactor: extract user service
