+++ backend/src/controllers/usersController.js
@@ -1,39 +1,10 @@
-const fs = require('fs').promises;
-const path = require('path');
 const jwt = require('jsonwebtoken');
-
-const USERS_FILE = path.join(__dirname, '..', 'users.json');
-
-let users = [];
-
-async function loadUsers() {
-  try {
-    const data = await fs.readFile(USERS_FILE, 'utf-8');
-    users = JSON.parse(data);
-  } catch (error) {
-    if (error.code !== 'ENOENT') {
-      console.error('Failed to load users:', error);
-    }
-  }
-}
-
-async function saveUsers() {
-  try {
-    await fs.writeFile(USERS_FILE, JSON.stringify(users, null, 2));
-  } catch (error) {
-    console.error('Failed to save users:', error);
-    throw error;
-  }
-}
-
-loadUsers();
+const usersService = require('../services/usersService');
 
 async function createUser(req, res) {
   const { name, email, phone, password } = req.body;
-  const user = { name, email, phone, password };
-  users.push(user);
   try {
-    await saveUsers();
+    const user = await usersService.createUser({ name, email, phone, password });
     res.status(201).json(user);
   } catch (error) {
     res.status(500).json({ error: 'Could not save user' });
@@ -41,11 +12,13 @@ async function createUser(req, res) {
 }
 
 function getUsers(req, res) {
+  const users = usersService.getUsers();
   res.json(users);
 }
 
 function loginUser(req, res) {
   const { email, password } = req.body;
+  const users = usersService.getUsers();
   const user = users.find((u) => u.email === email && u.password === password);
   if (!user) {
     return res.status(401).json({ error: 'Invalid credentials' });
@@ -55,4 +28,3 @@ function loginUser(req, res) {
 }
 
 module.exports = { createUser, getUsers, loginUser };
-
